bump to go1.18
Added go.work file to make gopls work. The go.work file should
generally not be checked into git with the exception if the
repository contains multiple modules which depend on each other.
The example in cmd/modprobe is a separate module to minimize the
number of default packages of kmod.